Changed top package name to io.curity.
