+++ oauthtest-core/build.gradle
@@ -8,7 +8,7 @@ repositories {
     jcenter()
 }
 
-mainClassName = "com.athaydes.oauth.core.OAuthTestApplication"
+mainClassName = "io.curity.oauth.core.OAuthTestApplication"
 
 dependencies {
     compileOnly "com.google.code.findbugs:annotations:3.0.1"
