--- oauthtest-core/src/main/java/com/athaydes/oauth/core/util/event/Event.java
@@ -1,8 +1,0 @@
-package com.athaydes.oauth.core.util.event;
-
-public interface Event {
-
-
-
-
-}
