--- oauthtest-core/src/main/java/com/athaydes/oauth/core/util/event/Subscriber.java
@@ -1,8 +1,0 @@
-package com.athaydes.oauth.core.util.event;
-
-@FunctionalInterface
-public interface Subscriber<E extends Event> {
-
-    void handle( E event );
-
-}
