+++ oauthtest-core/src/main/java/io/curity/oauth/core/request/HttpResponseEvent.java
@@ -1,0 +1,20 @@
+package io.curity.oauth.core.request;
+
+import io.curity.oauth.core.util.event.Event;
+import com.sun.jersey.api.client.ClientResponse;
+
+/**
+ * Simple ClientResponse Wrapper class that allows responses to be published on the EventBus.
+ */
+public class HttpResponseEvent implements Event {
+
+    private final ClientResponse response;
+
+    public HttpResponseEvent( ClientResponse response ) {
+        this.response = response;
+    }
+
+    public ClientResponse getResponse() {
+        return response;
+    }
+}
