+++ oauthtest-core/src/main/java/io/curity/oauth/core/util/ListUtils.java
@@ -1,0 +1,18 @@
+package io.curity.oauth.core.util;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ *
+ */
+public class ListUtils {
+
+    public static <T> List<T> append( List<T> a, List<T> b ) {
+        List<T> result = new ArrayList<>( a.size() + b.size() );
+        result.addAll( a );
+        result.addAll( b );
+        return result;
+    }
+
+}
