+++ oauthtest-core/src/main/java/io/curity/oauth/core/util/event/Event.java
@@ -1,0 +1,8 @@
+package io.curity.oauth.core.util.event;
+
+public interface Event {
+
+
+
+
+}
