+++ oauthtest-core/src/main/java/io/curity/oauth/core/util/event/Subscriber.java
@@ -1,0 +1,8 @@
+package io.curity.oauth.core.util.event;
+
+@FunctionalInterface
+public interface Subscriber<E extends Event> {
+
+    void handle( E event );
+
+}
