+++ oauthtest-core/src/main/resources/fxml/arrows.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.shape.SVGPath?>
 <?import javafx.scene.text.Text?>
 <VBox spacing="5.0" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1"
-      fx:controller="com.athaydes.oauth.core.component.Arrows">
+      fx:controller="io.curity.oauth.core.component.Arrows">
     <children>
         <Text fx:id="currentStep" text="No Steps" textAlignment="CENTER"/>
         <HBox maxHeight="-Infinity" maxWidth="-Infinity" spacing="10.0">
