+++ oauthtest-core/src/main/resources/fxml/code-flow-authz-req.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.layout.RowConstraints?>
 <GridPane alignment="CENTER" hgap="5.0" vgap="5.0" xmlns="http://javafx.com/javafx/8.0.111"
           xmlns:fx="http://javafx.com/fxml/1"
-          fx:controller="com.athaydes.oauth.core.controller.view.CodeFlowAuthzRequestView">
+          fx:controller="io.curity.oauth.core.controller.view.CodeFlowAuthzRequestView">
     <columnConstraints>
         <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0"/>
         <ColumnConstraints hgrow="SOMETIMES" minWidth="50.0" prefWidth="250.0"/>
